+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/gitlab
 
+## 0.16.12
+
+### Patch Changes
+
+- Updated dependencies [11dae5b]
+  - @cat-factory/contracts@0.259.0
+  - @cat-factory/kernel@0.257.0
+
 ## 0.16.11
 
 ### Patch Changes